# backend/routes/swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "Defines API endpoints for the FieldDay mark/recapture application.",
      "version": "1.0.1",
      "title": "FieldDay - Web Server",
      "termsOfService": "https://www.asu.edu/aad/manuals/acd/acd125.html",
      "contact": {
        "email": "krshelle@asu.edu"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      }
    },
    "host": "fhqtkwwgl1.execute-api.us-west-1.amazonaws.com",
    "basePath": "/dev",
...
    "schemes": [
      "https",
      "http"
    ],
    "paths": {
...
      "/answer_set/{set_name}": {
        "get": {
          "tags": [
            "answer set"
          ],
          "summary": "Finds answer set by set name",
          "description": "",
          "operationId": "findAnswerSetBySetName",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "set_name",
              "type": "string",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "$ref": "#/definitions/AnswerSet"
              }
            },
            "401": {
              "description": "Not authorized"
            }
          },
          "security": [
            {
              "Bearer": []
            }
          ]
        },
        "put": {
          "tags": [
            "answer set"
          ],
          "summary": "Updated answer set",
          "description": "",
          "operationId": "updateAnswerSet",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "set_name",
              "type": "string",
              "required": true
            },
            {
              "in": "body",
              "name": "body",
              "description": "Updated answer set object",
              "required": true,
              "schema": {
                "$ref": "#/definitions/AnswerSet"
              }
            }
          ],
          "responses": {
            "400": {
              "description": "Invalid answer set name supplied"
            },
            "404": {
              "description": "Answer set not found"
            }
          },
          "security": [
            {
              "Bearer": []
            }
          ]
        },
        "delete": {
          "tags": [
            "answer set"
          ],
          "summary": "Deletes an answer set",
          "description": "",
          "operationId": "deleteAnswerSet",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "set_name",
              "type": "string",
              "required": true
            }
          ],
          "responses": {
            "400": {
              "description": "Invalid set name supplied"
            },
            "401": {
              "description": "Not authorized"
            },
            "404": {
              "description": "Answer set not found"
            }
          },
          "security": [
            {
              "Bearer": []
            }
          ]
        }
      },
...
      "/data_entry/{session_id}/{entry_id}": {
        "get": {
          "tags": [
            "data entry"
          ],
          "summary": "Finds data entry by session and entry id",
          "description": "",
          "operationId": "findDataEntryById",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "session_id",
              "type": "integer",
              "required": true
            },
            {
              "in": "path",
              "name": "entry_id",
              "type": "integer",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "$ref": "#/definitions/DataEntry"
              }
            },
            "400": {
              "description": "Invalid session or entry id"
            },
            "401": {
              "description": "Not authorized"
            }
          },
          "security": [
            {
              "Bearer": []
            }
          ]
        },
        "put": {
          "tags": [
            "data entry"
          ],
          "summary": "Updated data entry",
          "description": "",
          "operationId": "updateDataEntry",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "session_id",
              "type": "integer",
              "required": true
            },
            {
              "in": "path",
              "name": "entry_id",
              "type": "integer",
              "required": true
            },
            {
              "in": "body",
              "name": "body",
              "description": "Updated data entry",
              "required": true,
              "schema": {
                "$ref": "#/definitions/DataEntry"
              }
            }
          ],
          "responses": {
            "400": {
              "description": "Invalid session or entry id"
            },
            "401": {
              "description": "Not authorized"
            },
            "404": {
              "description": "Data Entry not found"
            }
          },
          "security": [
            {
              "Bearer": []
            }
          ]
        },
        "delete": {
          "tags": [
            "data entry"
          ],
          "summary": "Deletes a data entry set",
          "description": "",
          "operationId": "deleteDataEntry",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "session_id",
              "type": "integer",
              "required": true
            },
            {
              "in": "path",
              "name": "entry_id",
              "type": "integer",
              "required": true
            }
          ],
          "responses": {
            "400": {
              "description": "Invalid session or entry id supplied"
            },
            "401": {
              "description": "Not authorized"
            },
            "404": {
              "description": "Data entry not found"
            }
          },
          "security": [
            {
              "Bearer": []
            }
          ]
        }
      },
...
      "/session/{session_id}": {
        "get": {
          "tags": [
            "session"
          ],
          "summary": "Finds session by id",
          "description": "",
          "operationId": "findSessionById",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "session_id",
              "type": "integer",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "$ref": "#/definitions/Session"
              }
            },
            "400": {
              "description": "Invalid session id"
            },
            "401": {
              "description": "Not authorized"
            }
          },
          "security": [
            {
              "Bearer": []
            }
          ]
        },
        "put": {
          "tags": [
            "session"
          ],
          "summary": "Updated session",
          "description": "",
          "operationId": "updateSession",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "session_id",
              "type": "integer",
              "required": true
            },
            {
              "in": "body",
              "name": "body",
              "description": "Updated session",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Session"
              }
            }
          ],
          "responses": {
            "400": {
              "description": "Invalid session id"
            },
            "401": {
              "description": "Not authorized"
            },
            "404": {
              "description": "Session not found"
            }
          },
          "security": [
            {
              "Bearer": []
            }
          ]
        },
        "delete": {
          "tags": [
            "session"
          ],
          "summary": "Deletes a session",
          "description": "",
          "operationId": "deleteSession",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "session_id",
              "type": "integer",
              "required": true
            }
          ],
          "responses": {
            "400": {
              "description": "Invalid session id supplied"
            },
            "401": {
              "description": "Not authorized"
            },
            "404": {
              "description": "Session not found"
            }
          },
          "security": [
            {
              "Bearer": []
            }
          ]
        }
      },
...
    },
    "securityDefinitions": {
      "Bearer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    },
    "definitions": {
      "AnswerSet": {
        "type": "object",
        "properties": {
          "set_name": {
            "type": "string"
          },
          "secondary_keys": {
            "type": "string"
          },
          "answers": {
            "type": "integer",
            "format": "int32"
          },
          "date_modified": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "DataEntry": {
        "type": "object",
        "required": [
          "name",
          "photoUrls"
        ],
        "properties": {
          "form_id": {
            "type": "integer",
            "format": "int64"
          },
          "session_id": {
            "type": "integer",
            "format": "int64"
          },
          "entry_id": {
            "type": "integer",
            "format": "int64"
          },
          "date_modified": {
            "type": "integer",
            "format": "int64"
          },
          "project_id": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "Session": {
        "type": "object",
        "properties": {
          "session_id": {
            "type": "integer",
            "format": "int64"
          },
          "session_json": {
            "type": "string"
          },
          "date_modified": {
            "type": "integer",
            "format": "int64"
          },
          "form_id": {
            "type": "integer",
            "format": "int64"
          },
          "project_id": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    },
...
  }